+++ .gitignore
@@ -7,4 +7,4 @@ compiled
 /doc
 /node_modules
 
-
+/.vscode/
