+++ jest.config.cjs
@@ -14,5 +14,8 @@ module.exports = {
     }
   ]
   },
-  testMatch: ['**/?(*.)+(spec|test).[tj]s?(x)'],
+  testMatch: [
+    '**/__tests__/**/*.[jt]s?(x)',
+    '**/?(*.)+(spec|test).[tj]s?(x)'
+  ],
 };
