+++ package.json
@@ -2,7 +2,7 @@
   "name": "pie-slang",
   "version": "1.0.0",
   "description": "Implementation of Pie, following The Little Typer",
-  "author": "CrazyVerseTech",
+  "author": "Ding Feng, Fan Haolin, Zhang Qixiang",
   "license": "ISC",
   "dependencies": {
     "ts-pattern": "^5.3.1"
@@ -12,14 +12,19 @@
     "@babel/highlight": "^7.25.9",
     "@babel/preset-env": "^7.26.0",
     "@babel/preset-typescript": "^7.26.0",
+    "@eslint/js": "^9.15.0",
     "@types/estree": "^1.0.6",
     "@types/jest": "^29.5.14",
     "babel-jest": "^29.7.0",
+    "eslint": "^9.15.0",
+    "eslint-plugin-react": "^7.37.2",
+    "globals": "^15.12.0",
     "jest": "^29.7.0",
     "resolve-cwd": "^3.0.0",
     "to-fast-properties": "^4.0.0",
     "ts-jest": "^29.2.5",
-    "typescript": "^5.6.3"
+    "typescript": "^5.6.3",
+    "typescript-eslint": "^8.15.0"
   },
   "scripts": {
     "test": "jest"
