+++ src/interpreter/__tests__/test_normalization.ts
@@ -1,0 +1,128 @@
+/* 
+// Sample mock objects and environment setup
+const mockSymbol = Symbol('x'); // Replace with actual Symbol
+const mockValue = new ADD1('ZERO'); // Mock Value for testing
+const mockCore: Core = ['add1', 'zero']; // Mock Core expression
+const sampleEnv: Env = [[mockSymbol, mockValue]]; // Example Env setup
+const mockClosure = new FO_CLOS(sampleEnv, mockSymbol, mockCore); // Example closure
+const mockNeutral = new N_Var(mockSymbol); // Example Neutral expression
+const mockDelay = new DELAY(new Box(new DELAY_CLOS(sampleEnv, mockCore))); // Example DELAY
+
+describe('Normalization Tests', () => {
+
+  test('later should delay evaluation', () => {
+    const result = later(sampleEnv, mockCore);
+    expect(result).toBeInstanceOf(DELAY); // Ensure the result is a delayed evaluation
+  });
+
+  test('undelay should evaluate DELAY_CLOS', () => {
+    const delayClos = new DELAY_CLOS(sampleEnv, mockCore);
+    const result = undelay(delayClos);
+    expect(result).toEqual(valOf(sampleEnv, mockCore)); // Should evaluate to the value of the expression
+  });
+
+  test('now should force evaluation for DELAY', () => {
+    const result = now(mockDelay);
+    expect(result).toEqual(valOf(sampleEnv, mockCore)); // Should evaluate the delayed value
+  });
+
+  test('getCoreType should return the type of a Core expression', () => {
+    const expr: Core = ['add1', 'zero']; // Example Core expression
+    const result = getCoreType(expr);
+    expect(result).toBe('add1'); // Expect the type of the expression
+  });
+
+  test('PIType should return a PI type or Value', () => {
+    const args: [Symbol, Value][] = [[Symbol('x'), 'NAT']];
+    const ret: Value = "NAT";
+    const result = PIType(args, ret);
+    expect(result).toBeInstanceOf(PI); // Ensure it returns a PI type
+  });
+
+  test('doAp should apply a function to a value', () => {
+    const rator = new LAM(mockSymbol, mockClosure);
+    const rand = mockValue;
+    const result = doAp(rator, rand);
+    expect(result).toEqual(valOfClosure(mockClosure, rand)); // Should apply the function correctly
+  });
+
+  test('doWhichNat should handle case analysis on Nat type', () => {
+    const result = doWhichNat('ZERO', mockValue, mockValue, mockValue);
+    expect(result).toBe(mockValue); // Expect the base case to return 'b'
+  });
+
+  test('doCar should return the first element of a pair', () => {
+    const cons = new CONS(mockValue, mockValue);
+    const result = doCar(cons);
+    expect(result).toBe(mockValue); // First element of the pair
+  });
+
+  test('doIterNat should iterate over natural numbers', () => {
+    const result = doIterNat('ZERO', mockValue, mockValue, mockValue);
+    expect(result).toBe(mockValue); // Iteration base case
+  });
+
+  test('doCdr should return the second element of a pair', () => {
+    const cons = new CONS(mockValue, mockValue);
+    const result = doCdr(cons);
+    expect(result).toBe(mockValue); // Second element of the pair
+  });
+
+  test('doIndList should perform induction on lists', () => {
+    const result = doIndList('NIL', mockValue, mockValue, mockValue);
+    expect(result).toBe(mockValue); // Base case of the list
+  });
+
+  test('doHead should return the head of a vector', () => {
+    const vec = new VEC_CONS(mockValue, mockValue);
+    const result = doHead(vec);
+    expect(result).toBe(mockValue); // Head of the vector
+  });
+
+  test('doTail should return the tail of a vector', () => {
+    const vec = new VEC_CONS(mockValue, mockValue);
+    const result = doTail(vec);
+    expect(result).toBe(mockValue); // Tail of the vector
+  });
+
+  test('indVecStepType should return the step type of a vector induction', () => {
+    const result = indVecStepType(mockValue, mockValue);
+    expect(result).toBeInstanceOf(PI); // Should return a Pi type
+  });
+
+  test('doIndVec should perform induction on vectors', () => {
+    const result = doIndVec('ZERO', 'VECNIL', mockValue, mockValue, mockValue);
+    expect(result).toBe(mockValue); // Base case of vector induction
+  });
+
+  test('valOf should evaluate Core expressions', () => {
+    const result = new ADD1(now((valOf(sampleEnv, mockCore)! as ADD1).smaller));
+    expect(result).toEqual(mockValue); // Assuming mockCore evaluates to mockValue
+  });
+
+  test('valOfClosure should evaluate closures with free variables', () => {
+    const result = valOfClosure(mockClosure, mockValue);
+    expect(result).toEqual(valOf(extendEnv(sampleEnv, mockSymbol, mockValue), mockCore));
+  });
+
+  test('valInCtx should evaluate Core in context', () => {
+    const ctx: Ctx = [[mockSymbol, new Def(mockValue, mockValue)]]; // Example context
+    const result = valInCtx(ctx, mockCore);
+    expect(result).toEqual(valOf(ctxToEnv(ctx), mockCore));
+  });
+
+  test('read_back_context should serialize context correctly', () => {
+    const ctx: Ctx = [[mockSymbol, new Def(mockValue, mockValue)]]; // Example context
+    const result = readBackContext(ctx);
+    expect(result).toBeInstanceOf(Array); // Assuming serialization to array form
+  });
+
+  test('ReadBackType should convert values back into Core expressions', () => {
+    const ctx: Ctx = [[mockSymbol, new Def(mockValue, mockValue)]]; // Example context
+    const result = readBackType(ctx, mockValue);
+    expect(result).toBeInstanceOf(Array); // Assuming Core is serialized into array form
+  });
+});
+
+
+*/
